helper

+++ helper.c
@@ -637,7 +637,9 @@ void remove_dir(unsigned char *disk, struct ext2_inode *dir, char *path) {
                     || (dir_entry->file_type == EXT2_FT_SYMLINK)) { // file or link
                     remove_file_or_link(disk, entry_inode, entry_path);
                 } else if (dir_entry->file_type == EXT2_FT_DIR) { // directory
-                    remove_dir(disk, entry_inode, entry_path);
+                    if (get_inode_num(disk, entry_inode) == 2) { // avoid root
+                        remove_dir(disk, entry_inode, entry_path);
+                    }
                 }
 
                 /* Moving to the next directory */
@@ -664,7 +666,9 @@ void remove_dir(unsigned char *disk, struct ext2_inode *dir, char *path) {
                         || (dir_entry->file_type == EXT2_FT_SYMLINK)) { // file or link
                         remove_file_or_link(disk, entry_inode, entry_path);
                     } else if (dir_entry->file_type == EXT2_FT_DIR) { // directory
-                        remove_dir(disk, entry_inode, entry_path);
+                        if (get_inode_num(disk, entry_inode) == 2) { // avoid root
+                            remove_dir(disk, entry_inode, entry_path);
+                        }
                     }
 
                     /* Moving to the next directory */
